requirements/base.txt
python-dotenv==1.0.1
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
idna==3.10
sniffio==1.3.1
typing_extensions==4.12.2
exceptiongroup==1.2.2
h11==0.14.0
six==1.16.0
tzdata==2024.2